# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi Documentation",
        "description": "Документация для сервиса справочники",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/countries-lang": {
            "get": {
                "tags": [
                    "country-lang"
                ],
                "summary": "List of country lang",
                "description": "Get list of countries lang",
                "operationId": "getCountriesLang",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/country-lang": {
            "post": {
                "tags": [
                    "country-lang"
                ],
                "summary": "Create country lang",
                "description": "Create country lang",
                "operationId": "createCountryLang",
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "english"
                                    },
                                    "lang": {
                                        "type": "string",
                                        "example": "en"
                                    },
                                    "country_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        },
        "/api/country-lang/{id}": {
            "get": {
                "tags": [
                    "country-lang"
                ],
                "summary": "Get country lang",
                "description": "Get country lang",
                "operationId": "getCountryLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "put": {
                "tags": [
                    "country-lang"
                ],
                "summary": "Update country lang",
                "description": "Update country lang",
                "operationId": "updateCountryLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "lang": {
                                        "type": "string"
                                    },
                                    "country_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            },
            "delete": {
                "tags": [
                    "country-lang"
                ],
                "summary": "Delete country lang",
                "description": "Delete country lang",
                "operationId": "deleteCountryLang",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "country lang ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Something is wrong"
                    }
                }
            }
        }
    },
...
}